feat: 상세페이지 레디스에서 가져오게 수정

+++ server/src/main/java/com/project/dailyAuction/board/Dto/BoardDto.java
@@ -41,7 +41,7 @@ public class BoardDto {
         private int currentPrice;
         private LocalDateTime createdAt;
         private LocalDateTime finishedAt;
-        private List<Long> history;
+        private List<Integer> history;
         private int myPrice;
 
         public void updateMyPrice(int myPrice) {
